+++ wvg.py
@@ -5,7 +5,7 @@ class wvg:
     # Initializes WVG. Weights and Quota are optional and defaulted to 0 if not provided
     def __init__(self, num_players, weights = None, quota = None):
         self.num_players = num_players
-        self.weights = np.zeros((self.num_players)) if weights == None else weights
+        self.weights = np.zeros((self.num_players), dtype=int) if weights == None else weights
 
         # Make sure there are same number of weights as players
         if self.num_players != len(self.weights):
@@ -25,7 +25,6 @@ class wvg:
         return self.quota
 
     # Setters
-    # Do we need some kind of check to make sure players still == len(weights)?
     def set_num_players(self, num_players):
         self.num_players = num_players
         self.weights = np.zeros((num_players))
